--- actions/delete.go
@@ -1,17 +1,0 @@
-package action
-
-import (
-	"fmt"
-
-	//"github.com/lxc/lxd/client"
-	//"github.com/lxc/lxd/shared/api"
-)
-
-func Delete(args1 string) (string, error){
-	resp := "ok"
-	var err error
-
-	fmt.Println(args1)
-
-	return resp, err
-}
