--- master/w8a/action/create.go
@@ -1,17 +1,0 @@
-package action
-
-import (
-	"fmt"
-
-	//"github.com/lxc/lxd/client"
-	//"github.com/lxc/lxd/shared/api"
-)
-
-func Create(args1 string) (string, error){
-	fmt.Println(args1)
-
-	resp := "ok"
-	var err error
-
-	return resp, err
-}
