--- master/w8a/action/delete.go
@@ -1,17 +1,0 @@
-package action
-
-import (
-	"fmt"
-
-	//"github.com/lxc/lxd/client"
-	//"github.com/lxc/lxd/shared/api"
-)
-
-func Delete(args1 string) (string, error){
-	fmt.Println(args1)
-
-	resp := "ok"
-	var err error
-
-	return resp, err
-}
